backend/requirements.txt
aiohttp
alembic
asyncpg
brotli
fastapi
httpx
networkx
numpy
orjson
pydantic
pydantic-settings
pynacl
redis
rich
scipy
sqlalchemy
tenacity
typer
uvicorn
zstandard
google-genai
pytest
pytest-asyncio

